# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  AuthErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        example: invalid_request
        type: string
      errorMsg:
        example: invalid request
        type: string
      requestId:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  ConversationErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  FileListResponse:
    properties:
      quota:
        $ref: '#/definitions/StorageQuotaResponse'
      results:
        items:
          $ref: '#/definitions/FileObjectResponse'
        type: array
      total:
        type: integer
    required:
    - quota
    - results
    - total
    type: object
  FileListResponseDoc:
    properties:
      data:
        $ref: '#/definitions/FileListResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  FileObjectResponse:
    properties:
      chunkCount:
        type: integer
      createdAt:
        type: string
      detectedMIME:
        type: string
      embedError:
        type: string
      embedStatus:
        type: string
      expiresAt:
        type: string
        x-nullable: true
        x-omitempty: false
      extractStatus:
        type: string
      fileCategory:
        type: string
      fileID:
        type: string
      fileName:
        type: string
      lastAccessedAt:
        type: string
        x-nullable: true
        x-omitempty: false
      mimeType:
        type: string
      processingErrorCode:
        type: string
      processingErrorMessage:
        type: string
      processingReady:
        type: boolean
      processingStatus:
        type: string
      purpose:
        type: string
      ragOptOut:
        type: boolean
      sha256:
        type: string
      sizeBytes:
        type: integer
      status:
        type: string
      updatedAt:
        type: string
    required:
    - chunkCount
    - createdAt
    - detectedMIME
    - embedError
    - embedStatus
    - expiresAt
    - extractStatus
    - fileCategory
    - fileID
    - fileName
    - lastAccessedAt
    - mimeType
    - processingErrorCode
    - processingErrorMessage
    - processingReady
    - processingStatus
    - purpose
    - ragOptOut
    - sha256
    - sizeBytes
    - status
    - updatedAt
    type: object
...
  FileUploadResponse:
    properties:
      file:
        $ref: '#/definitions/FileObjectResponse'
      quota:
        $ref: '#/definitions/StorageQuotaResponse'
      reused:
        type: boolean
    required:
    - file
    - quota
    - reused
    type: object
...
  LoginOptionsResponse:
    properties:
      emailRegistrationEnabled:
        type: boolean
      emailVerificationEnabled:
        type: boolean
      turnstileEnabled:
        type: boolean
      turnstileSiteKey:
        type: string
    required:
    - emailRegistrationEnabled
    - emailVerificationEnabled
    - turnstileEnabled
    - turnstileSiteKey
    type: object
  LoginOptionsResponseDoc:
    properties:
      data:
        $ref: '#/definitions/LoginOptionsResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  LoginResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
        x-nullable: true
      refreshExpiresAt:
        type: string
        x-nullable: true
      sessionID:
        type: string
      twoFactorChallengeToken:
        type: string
      twoFactorRequired:
        type: boolean
    required:
    - twoFactorRequired
    type: object
  LoginResponseDoc:
    properties:
      data:
        $ref: '#/definitions/LoginResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  StorageQuotaResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      quotaBytes:
        type: integer
      reservedBytes:
        type: integer
      updatedAt:
        type: string
      usedBytes:
        type: integer
      userID:
        type: integer
    required:
    - createdAt
    - id
    - quotaBytes
    - reservedBytes
    - updatedAt
    - usedBytes
    - userID
    type: object
...
  UploadFileResponseDoc:
    properties:
      data:
        $ref: '#/definitions/FileUploadResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
info:
  contact: {}
  description: DEEIX Chat 后端 API 文档
  title: DEEIX Chat API
  version: "0.4.0"
paths:
...
  /auth/login-options:
    get:
      description: Returns current Sub2 email registration and Turnstile settings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginOptionsResponseDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      summary: Get Sub2 login options
      tags:
      - auth
...
  /auth/refresh:
    post:
      description: Rotates the DEEIX session using only the HttpOnly refresh cookie
        and returns no refresh token in JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponseDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      summary: Refresh the browser session
      tags:
      - auth
...
  /files:
    get:
      consumes:
      - application/json
      description: 查询当前用户上传的文件
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: q
        type: string
      - description: '筛选，支持单值或逗号分隔多值: image,document,spreadsheet,presentation,code,pdf,audio,video'
        in: query
        name: kind
        type: string
      - description: '排序: created|name|size|last_used'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/FileListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ConversationErrorDoc'
      security:
      - BearerAuth: []
      summary: 文件分页列表
      tags:
      - chat
    post:
      consumes:
      - multipart/form-data
      description: 上传对话附件文件，统一存储并扣减用户配额（默认100MB）
      parameters:
      - description: 文件用途
        in: formData
        name: purpose
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UploadFileResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ConversationErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ConversationErrorDoc'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/ConversationErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ConversationErrorDoc'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"